agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-sock>=0.7.0",
    "yt-dlp>=2025.10.14",
    "google-genai>=1.0.0",
    "pillow>=12.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.1",
    "types-pillow>=10.2.0.20240822",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "flask_sock"
ignore_missing_imports = true
